chore(checkoutv2): remove CheckoutHeader component
